Moved fs2.Stream.Handle to fs2.Handle
