+++ core/shared/src/main/scala/fs2/fs2.scala
@@ -20,7 +20,4 @@ package object fs2 {
    * Sinks are typically applied with the `to` operation on `Stream`.
    */
   type Sink[F[_],-I] = Pipe[F,I,Unit]
-
-  type AsyncStep[F[_],A] = ScopedFuture[F, Pull[F, Nothing, (Chunk[A], Stream.Handle[F,A])]]
-  type AsyncStep1[F[_],A] = ScopedFuture[F, Pull[F, Nothing, (Option[A], Stream.Handle[F,A])]]
 }
