+++ core/shared/src/main/scala/fs2/pipe2.scala
@@ -1,6 +1,5 @@
 package fs2
 
-import Stream.Handle
 import fs2.{Pull => P}
 import fs2.util.{Async,Free,Functor,Sub1}
